+++ CHANGELOG.md
@@ -3,9 +3,4 @@
 ## [2025/05/26]
 
 - rename to LLMBrix
-- import fixes
-- doc updates
-- changelog added
-- develop branch removed (not needed for early dev)
-- Python version requirement increased to 3.11
-- Pydantic added to deps
+- major refactor and reduction of functionality
