+++ llmbrix/msg/__init__.py
@@ -1,4 +1,5 @@
 from .assistant_msg import AssistantMsg
+from .developer_msg import DeveloperMsg
 from .msg import Msg
 from .system_msg import SystemMsg
 from .tool_msg import ToolMsg
