+++ llmbrix/msg/developer_msg.py
@@ -1,0 +1,6 @@
+from llmbrix.msg.msg import Msg
+
+
+class DeveloperMsg(Msg):
+    content: str
+    role: str = "developer"
